adds old tests to level 18
