--- archive/level17-Additions.lark
@@ -1,3 +1,0 @@
-
-command: print | ifs (elifs)* elses? | input | for_loop | assign_list | list_access_var | change_list_item| comment | assign
-
