+++ grammars/level18-Additions.lark
@@ -1,2 +1,9 @@
 // adds round brackets and changes ask to input
 
+command: print | ifs (elifs)* elses? | input | for_loop | assign_list | list_access_var | assign
+
+print: _PRINT (_SPACE* _LEFT_BRACKET _SPACE* (quoted_text | list_access | var_access | sum) (_SPACE (quoted_text | list_access | var_access | sum))* _RIGHT_BRACKET)?
+input: var _SPACE _IS _SPACE _INPUT _LEFT_BRACKET  (quoted_text | list_access | var_access | sum) (_SPACE (quoted_text | list_access | var_access | sum))*  _RIGHT_BRACKET
+
+// todo: some of these vars sould alse be var_access (things that do not define like the loop var)
+for_loop: _FOR _SPACE (NAME | var) _SPACE _IN _SPACE _RANGE _LEFT_BRACKET (NUMBER | var) (_COMMA _SPACE|_COMMA) (NUMBER | var) ((_COMMA _SPACE |_COMMA) (NUMBER | var))? _RIGHT_BRACKET _COLON _EOL (_SPACE command) (_EOL _SPACE command)* _EOL _END_BLOCK
